requirements.txt
contourpy==1.0.5
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.37.4
greenlet==1.1.3
kiwisolver==1.4.4
matplotlib==3.6.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
tqdm==4.64.1
